Feat: 강좌 상세 페이지 지도 api 추가
