+++ src/types/lecture.d.ts
@@ -2,23 +2,25 @@ interface CotentHeaderParam {
     [key: string]: string;
 }
 
-interface lectureInfo {
-    lectureId: number;
-    uid: string;
-    maxParticipants: number;
-    category: string;
-    bankName?: string;
-    accoutName?: string;
-    accountNumber?: string;
+interface LectureData {
+    create_id?: number;
+    creator: string;
+    max_participants: number;
+    category?: string;
+    bank_name?: string;
+    account_name?: string;
+    account_number?: string;
     price?: number;
     title: string;
     content?: string;
-    startDate: Date;
-    endDate: Date;
+    cycle?: string;
+    count?: number;
+    start_date: string;
+    end_date: string;
     region: string;
-    imageUrl?: string;
-    createDate: Date;
+    image_url?: string;
+    createdDate?: Date;
 }
 
-declare module 'lectureInfo';
+declare module 'LectureData';
 declare module 'CotentHeaderParam';
